Hashlink buildscript fixes.
The merge brings in a hashlink submodule pointed at the 1.12 release.
The version we were pointing to in the github actions script is an older
snapshot release, so that's been updated to match.

Hashlink's github actions are used to source the binaries now, and that
includes some minor filename and folder structure changes. Build.xml's
hashlink configuration has been restored to what it was before the merge
but modified to take the new folder structure into account.